Input value type fixed

+++ DOM-projects/Project 2/pro2.js
@@ -3,9 +3,9 @@ const form = document.querySelector("form");
 form.addEventListener("submit" , function(e){
     e.preventDefault();
 
-    const height = parseInt(document.querySelector("#height").value);
+    const height = parseFloat(document.querySelector("#height").value);
 
-    const weight = parseInt(document.querySelector("#weight").value);
+    const weight = parseFloat(document.querySelector("#weight").value);
     
     const body = document.querySelector("body");
